Move all unit tests under the correct packages
